+++ Model/AlertEvent.cs
@@ -2,6 +2,6 @@ public class AlertEvent
 {   
      public int Id {get; set;}
     public string Message{get; set;} = "";
-    public int SoundValue{get; set;}
+    public float SoundValue{get; set;}
     public DateTime Timestamp {get; set;}
 }
